+++ packages/enhanced/src/rspack.ts
@@ -1,1 +1,4 @@
-export { ModuleFederationPlugin } from '@module-federation/rspack';
+export {
+  ModuleFederationPlugin,
+  PLUGIN_NAME,
+} from '@module-federation/rspack/plugin';