+++ packages/enhanced/src/webpack.ts
@@ -1,3 +1,6 @@
-import { default as ModuleFederationPlugin } from './wrapper/ModuleFederationPlugin';
+import {
+  default as ModuleFederationPlugin,
+  PLUGIN_NAME,
+} from './wrapper/ModuleFederationPlugin';
 
-export { ModuleFederationPlugin };
+export { ModuleFederationPlugin, PLUGIN_NAME };