+++ packages/runtime-core/__tests__/resources/app2/federation-remote-entry.js
@@ -1,0 +1,11 @@
+globalThis[`__FEDERATION_${'@federation-test/app2:custom'}__`] = {
+  get(scope) {
+    const moduleMap = {
+      './say'() {
+        return () => 'hello app2';
+      },
+    };
+    return moduleMap[scope];
+  },
+  init() {},
+};