+++ packages/modernjs/CHANGELOG.md
@@ -1,5 +1,54 @@
 # @module-federation/modern-js
 
+## 0.8.7
+
+### Patch Changes
+
+- 5f67582: chore(modern-js-plugin): add ssr option
+- Updated dependencies [835b09c]
+- Updated dependencies [f573ad0]
+- Updated dependencies [336f3d8]
+- Updated dependencies [4fd33fb]
+  - @module-federation/sdk@0.8.7
+  - @module-federation/enhanced@0.8.7
+  - @module-federation/node@2.6.20
+  - @module-federation/rsbuild-plugin@0.8.7
+
+## 0.8.6
+
+### Patch Changes
+
+- Updated dependencies [a1d46b7]
+  - @module-federation/rsbuild-plugin@0.8.6
+  - @module-federation/enhanced@0.8.6
+  - @module-federation/node@2.6.19
+  - @module-federation/sdk@0.8.6
+
+## 0.8.5
+
+### Patch Changes
+
+- @module-federation/enhanced@0.8.5
+- @module-federation/sdk@0.8.5
+- @module-federation/node@2.6.18
+
+## 0.8.4
+
+### Patch Changes
+
+- @module-federation/enhanced@0.8.4
+- @module-federation/node@2.6.17
+- @module-federation/sdk@0.8.4
+
+## 0.8.3
+
+### Patch Changes
+
+- Updated dependencies [8e172c8]
+  - @module-federation/sdk@0.8.3
+  - @module-federation/node@2.6.16
+  - @module-federation/enhanced@0.8.3
+
 ## 0.8.2
 
 ### Patch Changes