+++ packages/rsbuild-plugin/CHANGELOG.md
@@ -1,5 +1,46 @@
 # @module-federation/rsbuild-plugin
 
+## 0.8.7
+
+### Patch Changes
+
+- Updated dependencies [835b09c]
+- Updated dependencies [f573ad0]
+- Updated dependencies [336f3d8]
+- Updated dependencies [4fd33fb]
+  - @module-federation/sdk@0.8.7
+  - @module-federation/enhanced@0.8.7
+
+## 0.8.6
+
+### Patch Changes
+
+- a1d46b7: chore(rsbuild-plugin): split setUp function to help extend
+  - @module-federation/enhanced@0.8.6
+  - @module-federation/sdk@0.8.6
+
+## 0.8.5
+
+### Patch Changes
+
+- @module-federation/enhanced@0.8.5
+- @module-federation/sdk@0.8.5
+
+## 0.8.4
+
+### Patch Changes
+
+- @module-federation/enhanced@0.8.4
+- @module-federation/sdk@0.8.4
+
+## 0.8.3
+
+### Patch Changes
+
+- Updated dependencies [8e172c8]
+  - @module-federation/sdk@0.8.3
+  - @module-federation/enhanced@0.8.3
+
 ## 0.8.2
 
 ### Patch Changes